Add pass-ssh to the list
This adds my add-on "pass-ssh" to the list. I know that there's an existing add-on with this name but I think mine deserves the name more since it actually creates the ssh session instead of acting like "pass ssh-keygen". 

Not that it really matters since users can easily rename the add-on anyway.